Lots of changes - read desc
Added WOFF2 format

Fixed translation icon from 大 to 文

Added icons:

clip
speech-bubble-add
star-half
triangle-up
triangle-right
triangle-down

Added aliases:

gear a.k.a. cog
envelope a.k.a. message
ribbon a.k.a. ribbon-prize
star-circle a.k.a. beta
rocket a.k.a. beta-launch
bell-ring a.k.a. bell-on
x-circle a.k.a. close-input
x a.k.a. cross
stripe-compact a.k.a. stripe-menu
stripe-expanded a.k.a. stripe-flow
user-smile a.k.a. user-active
user-sleep a.k.a. user-inactive
pencil-lock-full a.k.a. edit-lock
speech-bubble-smile a.k.a. user-talk
image-lock-full a.k.a. image-lock
gallery a.k.a. photo-gallery
news a.k.a. newspaper
list-numbered a.k.a. list-sorted / list-ordered
list-bullet a.k.a. list-unsorted / list-unordered
edit a.k.a. pencil
edit-lock-1  a.k.a. pencil-lock-pt1
edit-lock-2  a.k.a. pencil-lock-pt2
edit-undo a.k.a. pencil-revert-full
edit-undo-1  a.k.a. pencil-revert-pt1
edit-undo-2  a.k.a. pencil-revert-pt2
map-pin-add-pt1 a.k.a. map-pin-add-1
map-pin-add-pt2 a.k.a. map-pin-add-2
template-add-pt1 a.k.a. template-add-1
template-add-pt2 a.k.a. template-add-2
eye-lid a.k.a. eye-closed
trash-slash a.k.a. trash-undo
block-slash a.k.a. block-undo